Develop

See merge request s11-bigdata-dist-sub1/S11P21D108!1

+++ total-repo/server/src/main/java/com/d108/project/domain/forum/board/repository/BoardRepository.java
@@ -1,7 +1,7 @@
 package com.d108.project.domain.forum.board.repository;
 
-import com.d108.project.domain.forum.board.domain.Board;
+import com.d108.project.domain.forum.board.entity.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BoardRepository extends JpaRepository<Board, Integer> {
+public interface BoardRepository extends JpaRepository<Board, Long> {
 }
