+++ total-repo/server/src/main/java/com/d108/project/domain/forum/post/dto/PostDeleteDto.java
@@ -4,5 +4,5 @@ import lombok.Getter;
 
 @Getter
 public class PostDeleteDto {
-    private Integer memberId;
+    private Long memberId;
 }
