+++ total-repo/server/src/main/java/com/d108/project/domain/forum/post/dto/PostResponseDto.java
@@ -1,14 +1,14 @@
 package com.d108.project.domain.forum.post.dto;
 
-import com.d108.project.domain.forum.post.domain.Post;
+import com.d108.project.domain.forum.post.entity.Post;
 import lombok.Builder;
 
 @Builder
 public class PostResponseDto {
-    private Integer id;
+    private Long id;
     private String title;
     private String content;
-    private Integer view;
+    private Long view;
 
     public static PostResponseDto from(Post post) {
         return PostResponseDto.builder()
