+++ total-repo/server/src/main/java/com/d108/project/domain/member/repository/MemberRepository.java
@@ -1,7 +1,7 @@
 package com.d108.project.domain.member.repository;
 
-import com.d108.project.domain.member.domain.Member;
+import com.d108.project.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepository extends JpaRepository<Member, Integer> {
+public interface MemberRepository extends JpaRepository<Member, Long> {
 }
